+++ plugins/ucc_admin/lib/ucc_admin_web/channels/admin_channel.ex
@@ -2,9 +2,11 @@ defmodule UccAdminWeb.AdminChannel do
 
   import Rebel.Query, warn: false
   import Rebel.Core, warn: false
+  import UcxUccWeb.Gettext
 
   alias UccAdminWeb.AdminView
   alias UccChatWeb.RebelChannel.{SideNav}
+  alias Rebel.SweetAlert
 
   require Logger
 
@@ -37,4 +39,38 @@ defmodule UccAdminWeb.AdminChannel do
     Phoenix.View.render_to_string AdminView, templ, bindings
   end
 
+  def admin_restart_server(socket, _sender) do
+    SweetAlert.swal_modal socket, ~g(Are you sure?),
+      ~g(This will disrupt servic for all active users), "warning",
+      [
+        showCancelButton: true, closeOnConfirm: false, closeOnCancel: true,
+        confirmButtonColor: "#DD6B55", confirmButtonText: ~g(Yes, restart it)
+      ],
+      confirm: fn _ ->
+        {title, message, status} =
+          case Application.get_env(:ucx_ucc, :restart_command) do
+            [command | args] ->
+              if System.find_executable(command) do
+                try do
+                  case System.cmd command, args do
+                    {_, 0} ->
+                      {~g"Restarting!", ~g"The server is being restarted!", "success"}
+                    {error, code} ->
+                      {gettext("Error %{code}", code: code), error , "error"}
+                  end
+                rescue
+                  _ ->
+                   {~g(Sorry), ~g(Something went wong), "error"}
+                end
+              else
+                {~g(Sorry), ~g(The configured restart command cannot be found), "error"}
+              end
+            nil ->
+              {~g(Sorry), ~g(The restart command is not configured!), "error"}
+
+          end
+        SweetAlert.swal(socket, title, message, status, timer: 5000, showConfirmButton: false)
+      end
+    socket
+  end
 end